[Fix #292] 검색기능 수정
