+++ app/(withGlobalLayout)/main/page.tsx
@@ -1,11 +1,9 @@
 "use client";
-import { useRef, useState } from "react";
+import { useEffect, useRef } from "react";
 import Image from "next/image";
 import VisitEmptyImage from "@/assets/images/visit_empty.svg";
 import S from "@/style/mainpage/main.module.css";
-import { toast } from "react-toastify";
 import GoTopButton from "@/utils/GoTopButton";
-import { getRoomsWithKeyword } from "@/utils/supabase/roomAPI";
 import MainCreateRoom from "@/components/main/CreateRoomModal";
 import { useCreateStore } from "@/store/toggle-store";
 import MainVisual from "@/components/main/MainVisual";
@@ -16,12 +14,13 @@ import MainSkeleton from "@/components/main/MainSkeleton";
 import useJoinRoom from "@/hooks/useJoinRoom";
 
 const Mainpage = () => {
-  const { rooms, setRooms } = useGetRoomsSocket();
+  const { rooms } = useGetRoomsSocket();
   const { isCreate, setIsCreate } = useCreateStore();
-  const [search, setSearch] = useState("");
+
   const isGoInClick = useRef(false);
   const { joinRoomHandler, fastJoinRoomHandler, gameStartHandler } = useJoinRoom();
 
+  console.log("메인페이지 룸리스트", rooms);
   //NOTE - 이전 코드
   /*
   useEffect(() => {
@@ -42,19 +41,6 @@ const Mainpage = () => {
   }, []);
   */
 
-  //NOTE - 방 목록 검색
-  const searchHandler = async (e: React.FormEvent<HTMLFormElement>) => {
-    e.preventDefault();
-    if (!search.trim()) return;
-
-    try {
-      const rooms = await getRoomsWithKeyword(search);
-      setRooms(rooms);
-    } catch (error) {
-      toast.error("검색 중 오류가 발생했습니다.");
-    }
-  };
-
   //NOTE - 방 목록 리스트 없을때 불러오는 로딩 모션
   if (!rooms) return <MainSkeleton />;
 
@@ -68,7 +54,7 @@ const Mainpage = () => {
           <div className={S.MainGnb}>
             <p>현재 활성화 되어있는 방</p>
             <div className={S.roomSearchAndButton}>
-              <RoomSearch searchHandler={searchHandler} search={search} setSearch={setSearch} />
+              <RoomSearch />
               <div className={S.gameGoButton}>
                 <button disabled={isGoInClick.current} onClick={fastJoinRoomHandler}>
                   빠른입장
